+++ noodles-cram/src/io/reader/header/container.rs
@@ -1,6 +1,8 @@
+//! CRAM header container reader.
+
 mod block;
 mod header;
-mod sam_header;
+pub mod sam_header;
 
 use std::io::{self, Read, Take};
 
@@ -9,7 +11,8 @@ use byteorder::{LittleEndian, ReadBytesExt};
 use self::block::read_block;
 pub(super) use self::header::read_header;
 
-pub(super) struct Reader<R> {
+/// A CRAM header container reader.
+pub struct Reader<R> {
     inner: Take<R>,
 }
 
@@ -23,9 +26,11 @@ where
         }
     }
 
-    pub(super) fn raw_sam_header_reader(
-        &mut self,
-    ) -> io::Result<sam_header::Reader<impl Read + '_>> {
+    /// Returns a raw SAM header reader.
+    ///
+    /// The caller is responsible of discarding any extra padding in the header text, e.g., using
+    /// [`sam_header::Reader::discard_to_end`].
+    pub fn raw_sam_header_reader(&mut self) -> io::Result<sam_header::Reader<impl Read + '_>> {
         let mut reader = read_block(&mut self.inner)?;
 
         let len = reader.read_i32::<LittleEndian>().and_then(|n| {
@@ -35,7 +40,8 @@ where
         Ok(sam_header::Reader::new(reader, len))
     }
 
-    pub(super) fn discard_to_end(&mut self) -> io::Result<u64> {
+    /// Discards all input until EOF.
+    pub fn discard_to_end(&mut self) -> io::Result<u64> {
         io::copy(&mut self.inner, &mut io::sink())
     }
 }